Change AccountRequest invocation parameters
Currently, we pass params['text'] directly to AccountRequest. That poses
at the same time a security risk, and a responsability problem: it is
not to AccountRequest to be aware of the external parameters formats and
constraints, but only to its validity in order to accomplish the very
specific task of requesting an account.